Move auth layout to layout folder.
